fixed mono save bug

+++ IDIKWA-App/Models/Settings.cs
@@ -8,5 +8,5 @@ using System.Threading.Tasks;
 namespace IDIKWA_App
 {
     [Serializable]
-    public record Settings(int BitRate, float MasterVolume, Dictionary<string, float> Volumes, TimeSpan Duration, bool Mono, string OutputPath, List<string> RecordingDevices, int SampleRate, string Culture);
+    public record Settings(int BitRate, int MasterVolume, Dictionary<string, int> Volumes, TimeSpan Duration, bool Mono, string OutputPath, List<string> RecordingDevices, int SampleRate, string Culture);
 }
